Refactor project structure with modular step-specific scripts
