+++ 1-bootstrap/step1_bootstrap.py
