+++ 3-aggregation/step3_analyze.py
