+++ 3-aggregation/visualize.py
@@ -1,506 +1,32 @@
-from graphviz import Digraph
-import numpy as np
-import matplotlib.pyplot as plt
-import seaborn as sns
-import pandas as pd
+"""
+Backward Compatibility Module for TracerX Marker Selection Pipeline
 
+This module provides backward compatibility by importing from the new modular structure.
+All functions are now available through their respective specialized modules.
 
-def collapse_nodes(U, C, E, A, W, threshold=0.0, only_leaf=False):
-    print("Loading collapse nodes")
-    # generate the tree
-    tree = ModifyTree(E)
-    if not only_leaf:
-        # collapse the branches with 0 length
-        branch_remove_idx = []
-        for i in range(tree.N-1, -1, -1):
-            for j in range(tree.N-1, -1, -1):
-                if int(E[i, j]) == 1 and sum(W[j, :]) == 0:
-                    branch_remove_idx.append(j)
-        for node in branch_remove_idx:
-            target = tree.cp_tree[node]
-            U[:, target] += U[:, node]
-            tree.delete_node(node)
+DEPRECATED: This file is maintained for backward compatibility only.
+New code should import directly from the specific modules:
+- tree_operations: For tree manipulation and node collapsing
+- tree_rendering: For Graphviz visualization and tree utilities  
+- step3_visualization: For main visualization and analysis functions
+"""
 
-        # collapse the nodes with 0 frequency
-        freq_remove_idx = []
-        freq_leaf_remove_idx = []
-        for i in range(tree.N-1, -1, -1):
-            if i in branch_remove_idx:
-                continue
-            if tree.is_root(i):
-                continue
-            if np.mean(U[:, i]) <= threshold:
-                if tree.num_children(i) == 1:
-                    freq_remove_idx.append(i)
-                elif tree.is_leaf(i):
-                    freq_leaf_remove_idx.append(i)
-        print(freq_remove_idx)
-        for node in freq_remove_idx:
-            target = tree.tree[node][0]
-            parent = tree.cp_tree[node]
-            tree.delete_node(node)
-            W[target, :] += W[node, :]
-        for node in freq_leaf_remove_idx:
-            tree.delete_node(node)
-    else:
-        # collapse the branches with 0 length and the child of the branch doesn't belong to leaf nodes
-        branch_remove_idx = []
-        print("Collapsing branch with length 0")
-        for i in range(tree.N - 1, -1, -1):
-            for j in range(tree.N - 1, -1, -1):
-                if int(E[i, j]) == 1 and sum(W[j, :]) == 0 and not tree.is_leaf(j):
-                    branch_remove_idx.append(j)
-        for node in branch_remove_idx:
-            target = tree.cp_tree[node]
-            U[:, target] += U[:, node]
-            tree.delete_node(node)
+# Import all functions from the new modular structure for backward compatibility
+from tree_operations import collapse_nodes, ModifyTree
+from tree_rendering import (
+    render_tumor_tree, add_prefix_tree, df2dict, W2node_dict,
+    generate_tree, generate_cp, root_searching, E2tree, validate_sample_consistency
+)
+from step3_visualization import (
+    combine_existing_tree_and_frequency_plots, analyze_tree_distribution, 
+    plot_mut_profile_comparison
+)
 
-        # collapse the leaf nodes with 0 frequency
-        freq_remove_idx = []
-        freq_leaf_remove_idx = []
-        print('Collapsing leaf nodes with frequency 0')
-        for i in range(tree.N - 1, -1, -1):
-            if i in branch_remove_idx:
-                continue
-            if np.mean(U[:, i]) <= threshold and tree.is_leaf(i):
-                freq_leaf_remove_idx.append(i)
-        for node in freq_leaf_remove_idx:
-            tree.delete_node(node)
-        for i in range(tree.N - 1, -1, -1):
-            if tree.num_children(i) == 1:
-                freq_remove_idx.append(i)
-        for node in freq_remove_idx:
-            target = tree.tree[node][0]
-            parent = tree.cp_tree[node]
-            tree.delete_node(node)
-            W[target, :] += W[node, :]
-
-    # delete those nodes
-    remove_idx = branch_remove_idx + freq_remove_idx + freq_leaf_remove_idx
-    print('Nodes ', remove_idx, 'will be collapsed.')
-    U_new = np.delete(U, remove_idx, axis=1)
-    C_new = np.delete(C, remove_idx, axis=0)
-    A_new = np.delete(A, remove_idx, axis=0)
-    A_new = np.delete(A_new, remove_idx, axis=1)
-    E_new = np.delete(tree.E, remove_idx, axis=0)
-    E_new = np.delete(E_new, remove_idx, axis=1)
-    W_new = np.delete(W, remove_idx, axis=0)
-    print("collapse", U_new.shape, C_new.shape)
-    return U_new, C_new, E_new, A_new, W_new
-
-class ModifyTree:
-    def __init__(self, E):
-        self.cp_tree = {}
-        self.tree = {}
-        self.E = E
-        self.N = len(E)
-        for i in range(self.N - 1, -1, -1):
-            for j in range(self.N - 1, -1, -1):
-                if int(E[i, j]) == 1:
-                    self.cp_tree[j] = i
-                    if i not in self.tree.keys():
-                        self.tree[i] = [j]
-                    else:
-                        self.tree[i].append(j)
-
-    def delete_node(self, idx):
-        if self.is_root(idx):
-            if self.num_children(idx) > 1:
-                raise('Cannot delete root node with more than one child!')
-            child = self.tree[idx][0]
-            del self.cp_tree[child]
-            del self.tree[idx]
-        elif self.is_leaf(idx):
-            parent = self.cp_tree[idx]
-            del self.cp_tree[idx]
-            if self.num_children(parent) == 1:
-                del self.tree[parent]
-            else:
-                self.tree[parent].remove(idx)
-        else:
-            parent = self.cp_tree[idx]
-            children = self.tree[idx]
-            self.tree[parent].remove(idx)
-            del self.cp_tree[idx]
-            for child in children:
-                self.cp_tree[child] = parent
-                self.tree[parent].append(child)
-                self.E[parent, child] = 1
-            del self.tree[idx]
-
-
-    def is_leaf(self, idx):
-        if idx not in self.tree.keys():
-            return True
-        else:
-            return False
-
-    def is_root(self, idx):
-        if idx in self.tree.keys() and idx not in self.cp_tree.keys():
-            return True
-        else:
-            return False
-
-    def num_children(self, idx):
-        if self.is_leaf(idx):
-            return 0
-        else:
-            return len(self.tree[idx])
-
-
-def add_prefix_tree(mutation):
-    mutation_prefixed = {}
-    for node, mut_list in mutation.items():
-        mutation_prefixed[node] = ["mut_" + str(mut) for mut in mut_list]
-    return mutation_prefixed
-
-def render_tumor_tree(tree_structure, node_dict):
-    w = Digraph(format='png')
-    edge_idx = 0
-    root = root_searching(tree_structure)
-    for p, c_list in tree_structure.items():
-        p_list = ['normal'] if p == root else node_dict[p]
-        p_final = tuple(set(p_list))
-        ne = len(p_final)
-        for c in c_list:
-            edge_idx += 1
-            if c not in node_dict:
-                c_type = str(c)
-            else:
-                c_type = c
-            c_final = tuple(set(node_dict[c_type]))
-            ne_ = len(c_final)
-            if ne >= 10:
-                p_str = str(p) + ' ' + ' '.join((str(e) for e in p_final[:ne//3])) + '\n' + ' '.join((str(e) for e in p_final[ne//3:ne//3*2])) + '\n' + ' '.join((str(e) for e in p_final[ne//3*2:]))
-                if ne_ >= 10:
-                    c_str = str(c) + ' ' + ' '.join((str(e) for e in c_final[:ne_//3])) + '\n' + ' '.join((str(e) for e in c_final[ne_//3:ne_//3*2])) + '\n' + ' '.join((str(e) for e in c_final[ne_//3*2:]))
-                else:
-                    c_str = str(c) + ' ' + ' '.join((str(e) for e in c_final))
-            else:
-                p_str = str(p) + ' ' + ' '.join((str(e) for e in p_final))
-                if ne_ >= 10:
-                    c_str = str(c) + ' ' + ' '.join((str(e) for e in c_final[:ne_//3])) + '\n' + ' '.join((str(e) for e in c_final[ne_//3:ne_//3*2])) + '\n' + ' '.join((str(e) for e in c_final[ne_//3*2:]))
-                else:
-                    c_str = str(c) + ' ' + ' '.join((str(e) for e in c_final))
-            w.edge(p_str, c_str, "b" + str(edge_idx))
-    return w
-
-def df2dict(df):
-    idx2name = {}
-    for i in range(len(df)):
-        if isinstance(df.loc[i]["Gene"], str):
-            name = df.loc[i]["Gene"]
-        else:
-            name = str(df.loc[i]["Chromosome"]) + '_' + str(df.loc[i]["Genomic Position"])
-        idx2name.setdefault(i, name)
-    return idx2name
-
-def W2node_dict(W_node, idx2name=None):
-    node_dict = {}
-    N, m = W_node.shape
-    for i in range(N):
-        node_dict.setdefault(i, [])
-        for j in range(m):
-            if W_node[i, j] == 1:
-                if idx2name is not None:
-                    node_dict[i].append(idx2name[j])
-                else:
-                    node_dict[i].append(j)
-    return node_dict
-
-
-def generate_tree(cp_tree):
-    tree = {}
-    for child, parent in cp_tree.items():
-        if parent in tree.keys():
-            tree[parent].append(child)
-        else:
-            tree[parent] = [child]
-    return tree
-
-def generate_cp(tree):
-    return {c: p  for p in tree.keys() for c in tree[p]} # child: parent
-
-def root_searching(tree):  # O(depth of tree) <= O(k)
-    tree_cp = generate_cp(tree)
-    start_node = list(tree_cp.keys())[0]
-    iter_count = 0
-    while True:
-        iter_count += 1
-        start_node = tree_cp[start_node]
-        if start_node not in tree_cp.keys():
-            break
-        if iter_count >= 100:
-            print("The directed tree exists self-loop.")
-            return None
-    return start_node
-def E2tree(E):
-    tree = {}
-    N = E.shape[0]
-    for i in range(N):
-        for j in range(N):
-            if E[i, j] == 1:
-                tree.setdefault(i, [])
-                tree[i].append(j)
-    return tree
-
-def validate_sample_consistency(clonal_freq_data):
-    """
-    Validate that all nodes have consistent sample counts.
-    
-    Args:
-        clonal_freq_data: Dictionary of clonal frequency data
-        
-    Returns:
-        int: Number of samples, or -1 if inconsistent
-    """
-    sample_counts = []
-    for node, freqs in clonal_freq_data.items():
-        for freq_sample in freqs:
-            sample_counts.append(len(freq_sample))
-    
-    if len(set(sample_counts)) > 1:
-        print(f"Warning: Inconsistent sample counts across nodes: {set(sample_counts)}")
-        return -1
-    
-    return sample_counts[0] if sample_counts else 0
-
-def combine_existing_tree_and_frequency_plots(tree_png_path, freq_png_path, output_path, 
-                                           patient_num, tree_idx, freq, type_name, actual_num_samples):
-    """
-    Combine existing tree and frequency PNG files into a side-by-side visualization.
-    
-    Args:
-        tree_png_path: Path to the Graphviz tree PNG file
-        freq_png_path: Path to the matplotlib frequency PNG file
-        output_path: Path for the combined output file
-        patient_num: Patient identifier
-        tree_idx: Tree index for title
-        freq: Tree frequency for title
-        type_name: Analysis type
-        actual_num_samples: Number of samples
-    """
-    import matplotlib.image as mpimg
-    from pathlib import Path
-    
-    # Check if both input files exist
-    if not Path(tree_png_path).exists():
-        print(f"Warning: Tree PNG not found: {tree_png_path}")
-        return None
-    
-    if not Path(freq_png_path).exists():
-        print(f"Warning: Frequency PNG not found: {freq_png_path}")
-        return None
-    
-    try:
-        # Load the images
-        tree_img = mpimg.imread(tree_png_path)
-        freq_img = mpimg.imread(freq_png_path)
-        
-        # Create figure with side-by-side subplots
-        fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(20, 10))
-        
-        # Display tree image on the left
-        ax1.imshow(tree_img)
-        ax1.axis('off')
-        ax1.set_title('Phylogenetic Tree Structure', fontsize=14, fontweight='bold', pad=20)
-        
-        # Display frequency image on the right
-        ax2.imshow(freq_img)
-        ax2.axis('off')
-        ax2.set_title('Clonal Frequencies', fontsize=14, fontweight='bold', pad=20)
-        
-        # Add main title for the entire figure
-        fig.suptitle(f'{patient_num} - Tree {tree_idx} - Combined Analysis\n'
-                    f'Bootstrap frequency: {freq} | Samples: {actual_num_samples} | Analysis: {type_name}', 
-                    fontsize=16, fontweight='bold', y=0.95)
-        
-        # Adjust layout
-        plt.tight_layout()
-        plt.subplots_adjust(top=0.85)
-        
-        # Save the combined visualization
-        plt.savefig(output_path, dpi=300, bbox_inches='tight', facecolor='white')
-        plt.close()
-        
-        print(f"Saved combined visualization: {output_path}")
-        return output_path
-        
-    except Exception as e:
-        print(f"Error combining images: {e}")
-        return None
-
-
-def analyze_tree_distribution(tree_distribution, directory, patient_num, type, fig=False, 
-                            sample_prefix='Region', custom_sample_names=None):
-    """
-    Analyze tree distribution with multi-sample support and create combined tree-frequency visualizations.
-    
-    This function generates all tree visualizations in a subdirectory and identifies the best tree
-    (highest frequency) to copy to the main results directory.
-    
-    Args:
-        tree_distribution: Tree distribution data from aggregation
-        directory: Output directory for files
-        patient_num: Patient identifier
-        type: Analysis type (e.g., 'initial')
-        fig: Whether to generate figures
-        sample_prefix: Prefix for sample names (default: 'Region')
-        custom_sample_names: Optional list of custom sample names
-    """
-    if not fig:
-        return
-        
-    # Create subdirectory for all tree visualizations
-    all_trees_dir = directory / f'all_trees_{type}'
-    all_trees_dir.mkdir(exist_ok=True, parents=True)
-    
-    # Track the best tree (highest frequency)
-    best_tree_idx = None
-    best_frequency = 0
-    best_tree_combined_path = None
-    
-    for idx in range(len(tree_distribution['freq'])):
-        tree_structure = tree_distribution['tree_structure'][idx]
-        cp_tree = tree_distribution['cp_tree'][idx]
-        node_dict = tree_distribution['node_dict'][idx]
-        node_dict_name = tree_distribution['node_dict_name'][idx]
-        freq = tree_distribution['freq'][idx]
-        clonal_freq = tree_distribution['clonal_freq'][idx]
-
-        # Track the best tree (highest frequency)
-        if freq > best_frequency:
-            best_frequency = freq
-            best_tree_idx = idx
-
-        # Validate sample consistency and detect number of samples
-        num_samples = validate_sample_consistency(clonal_freq)
-        if num_samples <= 0:
-            print(f"Error: Invalid or inconsistent sample data for tree {idx}. Skipping visualization.")
-            continue
-        
-        print(f"Processing tree {idx} with {num_samples} samples (frequency: {freq})")
-
-        ## Multi-sample clonal prevalence processing
-        prev_mat = []
-        for node, freqs in clonal_freq.items():
-            for freq_sample in freqs:
-                # Dynamic multi-sample processing (following run_data_multi_sample.py pattern)
-                actual_num_samples = len(freq_sample)
-                for sample_idx in range(actual_num_samples):
-                    # Generate sample names
-                    if custom_sample_names and sample_idx < len(custom_sample_names):
-                        sample_name = custom_sample_names[sample_idx]
-                    else:
-                        sample_name = f'{sample_prefix}_{sample_idx}'
-                    
-                    prev_mat.append({
-                        'fraction': freq_sample[sample_idx], 
-                        'sample': sample_name, 
-                        'clone': node
-                    })
-        
-        df_prev = pd.DataFrame(prev_mat)
-        
-        # Step 1: Generate phylogenetic tree (Graphviz) - temporary for combining
-        g = render_tumor_tree(tree_structure, node_dict_name)
-        tree_filename = all_trees_dir / f'{patient_num}_tree_dist{idx}_{type}_temp'
-        g.render(filename=tree_filename, cleanup=True)
-        tree_png_path = f"{tree_filename}.png"
-
-        # Step 2: Generate frequency plot (matplotlib) - temporary for combining
-        plt.figure(figsize=(12, 8))  # Slightly taller for better readability in combined view
-        
-        # Use seaborn color palette for better multi-sample visualization
-        actual_num_samples = len(df_prev['sample'].unique())
-        colors = sns.color_palette("Set2", actual_num_samples)
-        
-        # Create the bar plot with improved aesthetics
-        sns.barplot(data=df_prev, x='clone', y='fraction', hue='sample', palette=colors)
-        
-        # Improved title with sample count information
-        plt.title(f'{patient_num}_tree_{idx}_freq{freq} ({actual_num_samples} samples)', 
-                 fontsize=14, fontweight='bold')
-        
-        # Better legend positioning for multiple samples
-        plt.legend(title='Sample', bbox_to_anchor=(1.05, 1), loc='upper left')
-        
-        # Improve axis labels
-        plt.xlabel('Clone', fontweight='bold', fontsize=12)
-        plt.ylabel('Clonal Frequency', fontweight='bold', fontsize=12)
-        
-        # Rotate x-axis labels if many clones
-        if len(df_prev['clone'].unique()) > 8:
-            plt.xticks(rotation=45)
-        
-        # Add grid for better readability
-        plt.grid(True, alpha=0.3, axis='y')
-        
-        # Save frequency plot temporarily
-        freq_filename = all_trees_dir / f'{patient_num}_freq_dist{idx}_{type}_temp.png'
-        plt.tight_layout()
-        plt.savefig(freq_filename, dpi=300, bbox_inches='tight')
-        plt.close()
-        
-        # Step 3: Combine both plots side-by-side and save to subdirectory
-        combined_filename = all_trees_dir / f'{patient_num}_combined_tree_freq_{idx}_{type}.png'
-        combined_result = combine_existing_tree_and_frequency_plots(
-            tree_png_path=tree_png_path,
-            freq_png_path=freq_filename,
-            output_path=combined_filename,
-            patient_num=patient_num,
-            tree_idx=idx,
-            freq=freq,
-            type_name=type,
-            actual_num_samples=actual_num_samples
-        )
-        
-        # Track the best tree's combined visualization path
-        if idx == best_tree_idx:
-            best_tree_combined_path = combined_filename
-        
-        # Clean up temporary files
-        import os
-        try:
-            if os.path.exists(tree_png_path):
-                os.remove(tree_png_path)
-            if os.path.exists(freq_filename):
-                os.remove(freq_filename)
-        except Exception as e:
-            print(f"Warning: Could not clean up temporary files: {e}")
-        
-        if combined_result:
-            print(f"Successfully created combined visualization for tree {idx}")
-        else:
-            print(f"Failed to create combined visualization for tree {idx}")
-        
-        print(f"Saved combined visualization for tree {idx} with {actual_num_samples} samples")
-
-    # Copy the best tree visualization to the main aggregation results directory
-    if best_tree_combined_path and best_tree_combined_path.exists():
-        best_tree_main_path = directory / f'{patient_num}_combined_best_tree_{type}.png'
-        import shutil
-        try:
-            shutil.copy2(best_tree_combined_path, best_tree_main_path)
-            print(f"\nBest tree (index {best_tree_idx}, frequency {best_frequency}) copied to main directory:")
-            print(f"  {best_tree_main_path}")
-            print(f"All tree visualizations available in: {all_trees_dir}")
-        except Exception as e:
-            print(f"Warning: Failed to copy best tree visualization to main directory: {e}")
-    else:
-        print(f"Warning: Best tree visualization not found for copying to main directory")
-
-
-def plot_mut_profile_comparison(tree_distribution, aggregated_results_file, directory, patient):
-    """
-    Plot mutation profile comparison (placeholder function).
-    
-    Args:
-        tree_distribution: Tree distribution data
-        aggregated_results_file: Path to aggregated results file
-        directory: Output directory
-        patient: Patient identifier
-    """
-    # Placeholder implementation
-    print(f"Mutation profile comparison for patient {patient} - functionality to be implemented")
-    pass
+# Re-export all functions to maintain compatibility
+__all__ = [
+    'collapse_nodes', 'ModifyTree',
+    'render_tumor_tree', 'add_prefix_tree', 'df2dict', 'W2node_dict',
+    'generate_tree', 'generate_cp', 'root_searching', 'E2tree', 'validate_sample_consistency',
+    'combine_existing_tree_and_frequency_plots', 'analyze_tree_distribution', 
+    'plot_mut_profile_comparison'
+]
