+++ 4-markers/step4_optimize_fraction.py
@@ -1,5 +1,5 @@
 import numpy as np
-from optimize import *
+from step4_optimize import *
 import gurobipy as gp
 import math
